K Cup Counter
This is a project I came up with to count how much K-Cuos are used in the office. I was curious to know how much waste is generated from using K-Cups.

People may not be comfortable knowing how many K-Cups are used because of fear of being tracked somehow. But this project is to bring awareness as to how much waste is being generated by using K-Cups.